# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Hospital Vision 2020 Conference 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions/lookup:
    get:
      operationId: lookupParticipantSubmissions
      tags: [files]
      summary: Lookup a participant by name for secure file submission
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching participants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipantLookupItem"
  /submissions/send-otp:
    post:
      operationId: sendSubmissionOTP
      tags: [files]
      summary: Request identity verification OTP for file submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOTPInput"
      responses:
        "200":
          description: OTP request successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessageResponse"
  /submissions/verify-otp:
    post:
      operationId: verifySubmissionOTP
      tags: [files]
      summary: Verify identity OTP and obtain auth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPInput"
      responses:
        "200":
          description: OTP verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # SYSTEM USERS (admin only)
...
components:
  schemas:
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/CurrentUser"
    CurrentUser:
      type: object
      required: [id, name, mobile, userType]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        userType:
          type: string
          enum: [participant, admin, super_admin, track_coordinator, food_coordinator, scientific_committee, pr_member, coordinator_view_only]
        assignedTrack:
          type: ["string", "null"]
        participantId:
          type: ["integer", "null"]
        sessionTimeoutMinutes:
          type: integer
        permissions:
          type: array
          items:
            type: string

    # PARTICIPANT SCHEMAS
...
    ParticipantLookupItem:
      type: object
      required: [id, name, registrationNumber, institution, maskedEmail, maskedMobile, roles, needsPPT, needsPoster, hasPassword]
      properties:
        id:
          type: integer
        name:
          type: string
        registrationNumber:
          type: string
        institution:
          type: string
        maskedEmail:
          type: string
        maskedMobile:
          type: string
        roles:
          type: array
          items:
            type: string
        needsPPT:
          type: boolean
        needsPoster:
          type: boolean
        hasPassword:
          type: boolean
    SendOTPInput:
      type: object
      required: [participantId, sendTo]
      properties:
        participantId:
          type: integer
        sendTo:
          type: string
        value:
          type: string
    VerifyOTPInput:
      type: object
      required: [participantId, otp]
      properties:
        participantId:
          type: integer
        otp:
          type: string
    SimpleMessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    # SYSTEM USER SCHEMAS
...